r/kubernetes_cluster: splitting the tests out
